Init Code
This will create a window that will be placed at the bottom corner of the primary monitor display.

This will have react to make a nice UI to hold the data within redux and make async calls to the API. Finally it will use IPC to write and get configs from the disk for the user's config for the different api endpoints to hit (for many API keys) or projects to poll in the update process.